Make CursorBuilder independent of mapping scope
